refactor: move ports directory from application to domain
